move certificate generation to runnables
