+++ src/main/lens-proxy/lens-proxy-certificate.injectable.ts
@@ -3,40 +3,21 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import { generate } from "selfsigned";
+import type { SelfSignedCert } from "selfsigned";
 import { lensProxyCertificateInjectionToken } from "../../common/certificate/lens-proxy-certificate-injection-token";
 
 const lensProxyCertificateInjectable = getInjectable({
   id: "lens-proxy-certificate",
   instantiate: () => {
-    const cert = generate([
-      { name: "commonName", value: "Lens Certificate Authority" },
-      { name: "organizationName", value: "Lens" },
-    ], {
-      keySize: 2048,
-      algorithm: "sha256",
-      days: 365,
-      extensions: [
-        {
-          name: "basicConstraints",
-          cA: true,
-        },
-        {
-          name: "subjectAltName",
-          altNames: [
-            { type: 2, value: "*.lens.app" },
-            { type: 2, value: "lens.app" },
-            { type: 2, value: "localhost" },
-            { type: 7, ip: "127.0.0.1" },
-          ],
-        },
-      ],
-    });
+    let certState: SelfSignedCert;
 
     return {
-      get: () => cert,
-      set: (): void => {
-        throw "cannot override cert";
+      get: () => certState,
+      set: (cert: SelfSignedCert): void => {
+        if (certState) {
+          throw "cannot override cert";
+        }
+        certState = cert;
       },
     };
   },
